requirements.txt
langchain
pandas==1.5.0
pydantic==1.10.8
snowflake_snowpark_python==1.6.1
snowflake-snowpark-python[pandas]
streamlit==1.24.0
pymilvus
grpcio<=1.56.0,>=1.49.1
unstructured==0.7.12
tiktoken==0.4.0
openai==0.27.8
black==23.3.0
replicate==0.8.4
langchain-experimental
sqlalchemy>2.0.0
snowflake-sqlalchemy==1.5.0